+++ database/migrations/2020_08_11_070230_create_detail_transactions_table.php
@@ -18,7 +18,7 @@ class CreateDetailTransactionsTable extends Migration
             $table->unsignedBigInteger('t_id');
             $table->unsignedBigInteger('p_id');
             $table->integer('qty'); 
-            $table->integer('sub_total');
+            $table->decimal('sub_total');
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP'));
